refactor: controls と単純ダイアログを component 配下へ移動する
- controls/ の 3 部品を component/controls/ へ、runtime domain 依存の
  ない 5 ダイアログ (Modal/EndMatch/Reset/Riichi/Ryukyoku) を
  component/dialogs/ へ git mv (内容はパス修正のみ)
- HanFuInput の inline type import を import type 形式へ正規化
- Ron/Tsumo ダイアログは domain 依存除去とあわせて後続で移行する
