+++ src/presentation/component/controls/CounterInput.tsx
