+++ src/presentation/component/controls/HanFuInput.tsx
@@ -1,4 +1,4 @@
-import { type Points } from '../../domain/movement';
+import type { Points } from '../../../domain/movement';
 
 const HAN_OPTIONS = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13] as const;
 const FU_OPTIONS = [20, 25, 30, 40, 50, 60, 70, 80, 90, 100, 110] as const;
